[FEAT] : 셀렉트 폴더, 타입 리스트 생성 및 영,한 셀렉트 선택 기능
